Push LSTM class into lstm_model module
